changed up casting to make IsAllowable() more usable

+++ maps/game_map.go
@@ -44,13 +44,13 @@ func (d sizes) IsUnlimited() bool {
 	return len(d) == 1 && d[0].Width == 0
 }
 
-func (d sizes) IsAllowable(Width uint, Height uint) bool {
+func (d sizes) IsAllowable(Width int, Height int) bool {
 	if d.IsUnlimited() {
 		return true
 	}
 
 	for _, size := range d {
-		if size.Width == Width && size.Height == Height {
+		if size.Width == uint(Width) && size.Height == uint(Height) {
 			return true
 		}
 	}
